requirements.txt
dotenv==0.9.9
fastapi==0.115.11
mangum==0.19.0
nltk==3.9.1
psycopg2-binary==2.9.10
uvicorn==0.34.0
